# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API Minha Folga
  version: 1.1.0
  summary: Cadastro de interesse, preferências, atendimento, ferramentas da Bia (Hal-AI) e administração.
  description: |
    Contrato versionado entre `frontend/` (Vue) e `backend/` (Node/Express). Todas as rotas vivem sob `/api`,
    na mesma origem do site. Regras gerais:

    - Erros sempre em JSON: `{ "error": { "code", "message", "fields"? , ...extra }, "requestId" }`.
      `message` é texto em português que pode ser exibido ao usuário; `code` é estável.
    - Chamadas do navegador com método inseguro exigem `Content-Type: application/json` e, se houver
      cabeçalho `Origin`, que ela esteja na lista permitida (defesa CSRF de base).
    - Corpos com campos inesperados são recusados (`validation_error`). Limite de corpo: 16 KB em todas as rotas.
    - Web Push usa autorização independente de navegador, endpoint cifrado e token técnico do dono; ver notifications.openapi.yaml.
    - Sessões usam cookies `HttpOnly; Secure; SameSite=Strict` definidos pelo servidor. Nenhum token vai para
      `localStorage`, query string ou pixels. Dados pessoais nunca aparecem em URL.
    - Respostas de cadastro e de pedido de saída são genéricas: não revelam se um número ou CPF já está cadastrado.
    - Em `PRE_LAUNCH`, toda rota `/api/credit/*` responde 403 `credit_phase_locked`.

    Decisão D1 (docs/DECISOES.md, 26/09/2026): este sistema **não mexe com WhatsApp**. Número oficial, Bia, webchat,
...
servers:
...
paths:
...
  /api/preferences/opt-out:
    post:
      tags: [ preferencias ]
      summary: Pedido de saída pelo número, sem sessão. Resposta sempre genérica; não mostra nada.
      description: |
        Se o número tiver cadastro, revoga o aviso de abertura e as novidades (estado `unsubscribed`), com trilha. Número sem
        cadastro recebe a mesma resposta. Anti-robô igual ao do cadastro (`antiBot`, campo-armadilha `website` → `202` sem
        gravar) e limite `waitlist.limits.optOutPerIpPerHour`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptOutRequest' }
      responses:
        '202': { description: Pedido registrado (resposta genérica)., content: { application/json: { schema: { $ref: '#/components/schemas/OptOutReceived' } } } }
        '400': { description: '`validation_error` (`phone` → "Confira o DDD e o número.") ou `form_expired`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/preferences/token:
    post:
      tags: [ preferencias ]
      summary: Troca o token do link seguro gerado pela Bia (issue_preferences_link; lido do fragmento da URL) por sessão.
      description: Uso único; validade de `waitlist.preferencesLinkTtlSeconds`. Não há acesso por código (D1).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false, required: [ token ], properties: { token: { type: string, minLength: 20, maxLength: 100 } } }
      responses:
        '200': { description: Sessão aberta (cookie mf_contact)., content: { application/json: { schema: { type: object, properties: { status: { const: ok } } } } } }
        '401': { description: '`invalid_or_expired_link`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/preferences:
    get:
      tags: [ preferencias ]
      summary: Lê as escolhas e os dados do titular, mascarados (somente com sessão válida).
      responses:
        '200': { description: Preferências., content: { application/json: { schema: { $ref: '#/components/schemas/Preferences' } } } }
        '401': { description: '`session_required` — orientar o link seguro pela Bia ou o pedido de saída, sem revelar dados.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [ preferencias ]
      summary: Altera finalidades, tema de interesse e dados gerais (faixas, cidade/UF, e-mail). Revogar `launch_notice` tira o contato da lista.
      description: CPF, telefone, nome e empregador não mudam por aqui (Bia com número atestado ou equipe de privacidade).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreferencesPatch' }
      responses:
        '200': { description: Preferências atualizadas., content: { application/json: { schema: { $ref: '#/components/schemas/Preferences' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { description: '`session_required`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: '`consent_version_outdated`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/preferences/deletion-request:
    post:
      tags: [ preferencias ]
      summary: Registra pedido de exclusão dos dados do titular (cumprido pela equipe de privacidade).
      responses:
        '202': { description: Pedido registrado., content: { application/json: { schema: { type: object, properties: { status: { const: received }, message: { type: string } } } } } }
        '401': { description: '`session_required`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/support:
    post:
      tags: [ atendimento ]
      summary: Registra solicitação de atendimento, gera protocolo aleatório e encaminha à fila interna.
      description: |
        A solicitação é persistida antes de qualquer notificação; falha de notificação não apaga o registro.
        O `trackingToken` é devolvido uma única vez e permite acompanhar o status; o protocolo sozinho não autoriza leitura.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupportRequest' }
      responses:
        '201': { description: Solicitação registrada., content: { application/json: { schema: { $ref: '#/components/schemas/SupportCreated' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { description: '`support_unavailable`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/support/{protocol}:
    get:
      tags: [ atendimento ]
      summary: Consulta o status de uma solicitação. Exige o token de acompanhamento do titular.
      parameters:
        - { in: path, name: protocol, required: true, schema: { type: string, pattern: '^MF-[0-9A-Z]{4}-[0-9A-Z]{4}-[0-9A-Z]{2}$' } }
        - { in: header, name: X-Support-Token, required: true, schema: { type: string } }
      responses:
        '200': { description: Status da solicitação., content: { application/json: { schema: { $ref: '#/components/schemas/SupportStatus' } } } }
        '404': { description: 'Protocolo inexistente ou token inválido (mesma resposta: `not_found`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { $ref: '#/components/responses/RateLimited' }

  # --------------------------------------------------------------------------------------------- métricas
...
  /api/admin/support:
    get:
      tags: [ admin ]
      summary: Fila de atendimento (support:read).
      description: Cada consulta é auditada (`admin.support_listed`, com filtros, página e quantidade; sem dados do titular).
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/SupportState' } }
        - { in: query, name: assignee, schema: { type: string, description: 'uuid, "me" ou "none"' } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200':
          description: Página da fila.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AdminSupportItem' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/support/{id}:
    get:
      tags: [ admin ]
      summary: Detalhe com mensagem e contato decifrados (support:read); acesso auditado.
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Detalhe., content: { application/json: { schema: { $ref: '#/components/schemas/AdminSupportDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [ admin ]
      summary: Altera status e/ou responsável (support:update). Registra responsável e horário.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                status: { $ref: '#/components/schemas/SupportState' }
                assigneeId: { type: [ string, 'null' ], format: uuid }
      responses:
        '200': { description: Atualizado., content: { application/json: { schema: { $ref: '#/components/schemas/AdminSupportDetail' } } } }
        '400': { description: '`invalid_transition` ou validação.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/support/{id}/notes:
    post:
      tags: [ admin ]
      summary: Registra nota interna ou resposta visível ao titular no acompanhamento (support:update).
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ note ]
              properties:
                note: { type: string, minLength: 1, maxLength: 2000 }
                visibleToRequester: { type: boolean, default: false }
      responses:
        '201': { description: Registrada., content: { application/json: { schema: { $ref: '#/components/schemas/AdminSupportDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/admin/users:
    get:
      tags: [ admin ]
      summary: Usuários do painel (users:manage).
      responses:
        '200': { description: Usuários., content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/AdminUser' } } } } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [ admin ]
      summary: Cria usuário individual com senha temporária exibida uma única vez (users:manage).
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ email, displayName, role ]
              properties:
                email: { type: string, format: email, maxLength: 200 }
                displayName: { type: string, minLength: 2, maxLength: 80 }
                role: { $ref: '#/components/schemas/AdminRole' }
      responses:
        '201':
          description: Criado.
          content:
            application/json:
              schema: { type: object, properties: { user: { $ref: '#/components/schemas/AdminUser' }, temporaryPassword: { type: string } } }
        '409': { description: '`email_in_use`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/admin/users/{id}:
    patch:
      tags: [ admin ]
      summary: Altera papel, desativa, redefine senha ou MFA (users:manage). Não permite remover o último admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                role: { $ref: '#/components/schemas/AdminRole' }
                disabled: { type: boolean }
                resetPassword: { type: boolean }
                resetMfa: { type: boolean }
      responses:
        '200':
          description: Atualizado.
          content:
            application/json:
              schema: { type: object, properties: { user: { $ref: '#/components/schemas/AdminUser' }, temporaryPassword: { type: string } } }
        '409': { description: '`last_admin`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
components:
  securitySchemes:
    adminSession: { type: apiKey, in: cookie, name: mf_admin }
    contactSession: { type: apiKey, in: cookie, name: mf_contact }
    agentApiKey: { type: apiKey, in: header, name: X-API-Key }
  parameters:
    Csrf:
      in: header
      name: X-CSRF-Token
      required: true
      description: Token da sessão administrativa (de /api/admin/auth/me ou do login). Obrigatório em métodos inseguros do admin.
      schema: { type: string }
  responses:
    ValidationError:
      description: '`validation_error` com `fields` { campo: mensagem }.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: '`rate_limited`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Forbidden:
      description: '`forbidden` (sem permissão; tentativa auditada).'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: '`not_found`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [ error ]
      properties:
        error:
          type: object
          required: [ code, message ]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            attemptsRemaining: { type: integer }
            retryAfterSeconds: { type: integer }
        requestId: { type: string }
...
    EmploymentType: { type: string, enum: [ clt, domestico, rural, outro ] }
    JobTenure: { type: string, enum: [ menos_3_meses, 3_a_12_meses, 1_a_3_anos, mais_3_anos ] }
    IncomeRange: { type: string, enum: [ ate_2000, 2000_a_4000, 4000_a_7000, acima_7000, nao_informar ] }
    Uf: { type: string, enum: [ AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN, RS, RO, RR, SC, SP, SE, TO ] }
    InterestTopic: { type: string, enum: [ organizar_compromissos, entender_alternativa, entender_portabilidade, so_conhecer ] }
    LeadState:
      type: string
      description: 'D1: received (site, telefone não validado) → verified (a Bia atestou o número) → invited; laterais unsubscribed e expired. verification_pending, waiting e human_support só em cadastros anteriores à D1.'
      enum: [ received, verified, invited, unsubscribed, expired, verification_pending, waiting, human_support ]
    SupportState: { type: string, enum: [ received, in_progress, answered, closed ] }
    AdminRole: { type: string, enum: [ admin, support, privacy, marketing ] }
...
    AntiBot:
      type: object
      additionalProperties: false
      required: [ token, nonce ]
      properties:
        token: { type: string, maxLength: 512, description: 'De GET /api/form-token.' }
        nonce: { oneOf: [ { type: string, pattern: '^\d{1,16}$' }, { type: integer, minimum: 0 } ], description: 'Solução da prova de trabalho.' }
...
    OptOutRequest:
      type: object
      additionalProperties: false
      required: [ phone, antiBot ]
      properties:
        phone: { type: string, maxLength: 20 }
        antiBot: { $ref: '#/components/schemas/AntiBot' }
        website: { type: string, description: 'Campo-armadilha.' }
    OptOutReceived:
      type: object
      required: [ status, message ]
      properties:
        status: { const: received }
        message: { type: string }
    Preferences:
      type: object
      required: [ preferredName, nameHint, contactHint, contactVerified, employmentType, state, purposes ]
      properties:
        preferredName: { type: string, description: Primeiro nome. }
        nameHint: { type: string, example: Ana P. S., description: 'Nome completo mascarado.' }
        contactHint: { type: string, example: '(11) •••••-••34' }
        contactVerified: { type: boolean, description: 'true depois que a própria pessoa falou com a Bia (número atestado).' }
        cpfHint: { type: [ string, 'null' ], example: '***.***.***-12' }
        emailHint: { type: [ string, 'null' ], example: 'a•••@e•••.com.br' }
        employerName: { type: [ string, 'null' ] }
        employmentType: { $ref: '#/components/schemas/EmploymentType' }
        jobTenure: { oneOf: [ { $ref: '#/components/schemas/JobTenure' }, { type: 'null' } ] }
        incomeRange: { oneOf: [ { $ref: '#/components/schemas/IncomeRange' }, { type: 'null' } ] }
        city: { type: [ string, 'null' ] }
        uf: { oneOf: [ { $ref: '#/components/schemas/Uf' }, { type: 'null' } ] }
        interestTopic: { oneOf: [ { $ref: '#/components/schemas/InterestTopic' }, { type: 'null' } ] }
        state: { $ref: '#/components/schemas/LeadState' }
        purposes:
          type: object
          properties:
            launch_notice: { $ref: '#/components/schemas/PurposeState' }
            marketing: { $ref: '#/components/schemas/PurposeState' }
    PurposeState:
      type: object
      properties:
        granted: { type: boolean }
        updatedAt: { type: string, format: date-time }
        consentTextVersion: { type: string }
    PreferencesPatch:
      type: object
      additionalProperties: false
      properties:
        purposes:
          type: object
          additionalProperties: false
          properties:
            launch_notice: { type: boolean }
            marketing: { type: boolean }
        consentVersions:
          type: object
          additionalProperties: false
          description: Obrigatório ao conceder (true) uma finalidade.
          properties:
            launch_notice: { type: string }
            marketing: { type: string }
        interestTopic: { oneOf: [ { $ref: '#/components/schemas/InterestTopic' }, { type: 'null' } ] }
        employmentType: { $ref: '#/components/schemas/EmploymentType' }
        jobTenure: { $ref: '#/components/schemas/JobTenure' }
        incomeRange: { $ref: '#/components/schemas/IncomeRange' }
        city: { type: string, maxLength: 80 }
        uf: { $ref: '#/components/schemas/Uf' }
        email: { type: [ string, 'null' ], maxLength: 160, description: 'null ou vazio remove o e-mail.' }
    SupportRequest:
      type: object
      additionalProperties: false
      required: [ name, contactMethod, contact, subject, message ]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        contactMethod: { type: string, enum: [ email, whatsapp ] }
        contact: { type: string, maxLength: 120, description: 'E-mail ("Confira o e-mail.") ou celular com DDD ("Confira o DDD e o número.").' }
        subject: { type: string, description: 'Uma de contracts/validation.json → support.subjects.' }
        message: { type: string, minLength: 10, maxLength: 2000 }
    SupportCreated:
      type: object
      required: [ protocol, trackingToken, status, message ]
      properties:
        protocol: { type: string, example: MF-7K3P-9QXA-2M }
        trackingToken: { type: string, description: Exibido uma única vez; compõe o link seguro de acompanhamento (fragmento da URL). }
        status: { const: received }
        message: { type: string }
    SupportStatus:
      type: object
      properties:
        protocol: { type: string }
        status: { $ref: '#/components/schemas/SupportState' }
        subject: { type: string }
        subjectLabel: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        timeline:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [ created, status_changed, reply ] }
              status: { $ref: '#/components/schemas/SupportState' }
              text: { type: [ string, 'null' ], description: Somente respostas marcadas como visíveis ao titular. }
              at: { type: string, format: date-time }
...
    AdminUserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/AdminRole' }
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/AdminRole' }
        mfaEnabled: { type: boolean }
        mustChangePassword: { type: boolean }
        disabled: { type: boolean }
        lastLoginAt: { type: [ string, 'null' ], format: date-time }
        createdAt: { type: string, format: date-time }
...
    AdminSupportItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        protocol: { type: string }
        subject: { type: string }
        status: { $ref: '#/components/schemas/SupportState' }
        source: { type: string, enum: [ web_form, agent ] }
        contactMethod: { type: string }
        contactHint: { type: string }
        assignee: { oneOf: [ { $ref: '#/components/schemas/AdminUserRef' }, { type: 'null' } ] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdminSupportDetail:
      allOf:
        - { $ref: '#/components/schemas/AdminSupportItem' }
        - type: object
          properties:
            requesterName: { type: string }
            contact: { type: string }
            message: { type: string }
            firstResponseAt: { type: [ string, 'null' ], format: date-time }
            closedAt: { type: [ string, 'null' ], format: date-time }
            events:
              type: array
              items:
                type: object
                properties:
                  type: { type: string }
                  fromStatus: { type: [ string, 'null' ] }
                  toStatus: { type: [ string, 'null' ] }
                  note: { type: [ string, 'null' ] }
                  visibleToRequester: { type: boolean }
                  actor: { oneOf: [ { $ref: '#/components/schemas/AdminUserRef' }, { type: 'null' } ] }
                  actorType: { type: string }
                  at: { type: string, format: date-time }
                  assignee:
                    description: Responsável naquele momento (eventos de atribuição).
                    oneOf:
                      - $ref: "#/components/schemas/AdminUserRef"
                      - type: "null"
...